Fix printing byte values in VK_EXT_host_image_copy array test logs
Printing uint8_t values with iostream treats them as characters rather
than numbers, so they must be cast to a larger integer type first.

Affects:
dEQP-VK.image.host_image_copy.array.*

Components: Vulkan

+++ external/vulkancts/modules/vulkan/image/vktImageHostImageCopyTests.cpp
@@ -3220,7 +3220,7 @@ tcu::TestStatus HostImageArrayCopyTestInstance::iterate(void)
             if (testData[i] != outputData[i])
             {
                 log << tcu::TestLog::Message << "At byte " << i << " generated data is " << testData[i]
-                    << ", but output data is " << outputData[i] << tcu::TestLog::EndMessage;
+                    << ", but output data is " << (int)outputData[i] << tcu::TestLog::EndMessage;
                 break;
             }
         }
